+++ test/eca/llm_providers/openai_test.clj
@@ -7,17 +7,15 @@
 (deftest ->messages-with-history-test
   (testing "no previous history"
     (is (match?
-         [{:role "user" :content "Hey"}]
-         (#'llm-providers.openai/->input-with-history [] "Hey"))))
+         []
+         (#'llm-providers.openai/past-messages->input []))))
   (testing "With basic text history"
     (is (match?
          [{:role "user" :content "Count with me: 1"}
-          {:role "assistant" :content "2"}
-          {:role "user" :content "3"}]
-         (#'llm-providers.openai/->input-with-history
+          {:role "assistant" :content "2"}]
+         (#'llm-providers.openai/past-messages->input
           [{:role "user" :content "Count with me: 1"}
-           {:role "assistant" :content "2"}]
-          "3"))))
+           {:role "assistant" :content "2"}]))))
   (testing "With tool_call history"
     (is (match?
          [{:role "user" :content "List the files you are allowed"}
@@ -29,9 +27,8 @@
           {:type "function_call_output"
            :call_id "call-1"
            :output "Allowed directories: /foo/bar\n"}
-          {:role "assistant" :content "I see /foo/bar"}
-          {:role "user" :content "Thanks"}]
-         (#'llm-providers.openai/->input-with-history
+          {:role "assistant" :content "I see /foo/bar"}]
+         (#'llm-providers.openai/past-messages->input
           [{:role "user" :content "List the files you are allowed"}
            {:role "assistant" :content "Ok!"}
            {:role "tool_call" :content {:id "call-1" :name "list_allowed_directories" :arguments {}}}
@@ -41,5 +38,4 @@
                                                :output {:contents [{:type :text
                                                                     :error false
                                                                     :content "Allowed directories: /foo/bar"}]}}}
-           {:role "assistant" :content "I see /foo/bar"}]
-          "Thanks")))))
+           {:role "assistant" :content "I see /foo/bar"}])))))